pyproject.toml
[tool.poetry]
name = "rest-api-demo"
version = "0.1.0"
description = ""
authors = ["Erik Tastepe <erik.tastepe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
flask-httpauth = "^4.8.0"
flask-restx = "^1.3.0"
flask-sqlalchemy = "^3.1.1"
pandas = "^2.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
